refactor(theme): pull default component configs into their own file
