+++ src/components/Theme/src/Theme.vue
@@ -13,7 +13,7 @@ import { showWarning } from '@square/maker/utils/debug';
 import key from './key';
 import defaultTheme from './default-theme';
 import { resolve, getPath } from './utils';
-import publicizeVars from './publicize-vars';
+import publicizeVars from './publicize-vars.cjs';
 
 // naive check, just checks for presence of a render field,
 // which is assumed to be a function, which is fine for now
